+++ console-frontend/src/plugin-sdk/nldd-entry.ts
@@ -1,4 +1,4 @@
-// Entry point for the NLDS plugin-UI bundle (public/plugin-ui/nldd.js). Importing
-// the package registers every <nldd-*> element as a side effect; plugin iframes
-// load the built IIFE via loadNlds() to render real NLDS components.
+// Entry point for the NLDD Design System plugin-UI bundle (public/plugin-ui/nldd.js).
+// Importing the package registers every <nldd-*> element as a side effect; plugin
+// iframes load the built IIFE via loadNldd() to render real components.
 import '@nldd/design-system';
