+++ plugins/openfsc/Dockerfile
@@ -1,6 +1,7 @@
 # Build the console UI (Vite, vanilla TS) → plugins/openfsc/console, which
-# console.go embeds via go:embed. NLDS is NOT bundled here: the app loads it at
-# runtime from the Console's shared /plugin-ui bundle. See docs/funs/FUN-18.adoc.
+# console.go embeds via go:embed. The NLDD Design System is NOT bundled here: the app
+# loads it at runtime from the Console's shared /plugin-ui bundle. See
+# docs/funs/FUN-18.adoc.
 FROM oven/bun:1-alpine AS console-ui
 WORKDIR /ui
 COPY plugins/openfsc/console-ui/package.json plugins/openfsc/console-ui/bun.lock ./
