+++ plugins/openfsc/console-ui/vite.config.ts
@@ -6,10 +6,10 @@ const root = fileURLToPath(new URL('.', import.meta.url));
 const entry = (name: string) => fileURLToPath(new URL(`${name}.html`, import.meta.url));
 
 // The plugin serves this app same-origin under /console/, and console.go's
-// `go:embed console/*` compiles the build output into the plugin binary. NLDS is
-// intentionally NOT bundled: the app uses <nldd-*> tags whose registrations come
-// from the shared /plugin-ui/nldd.js, loaded at runtime via loadNlds(). See
-// docs/funs/FUN-18.adoc.
+// `go:embed console/*` compiles the build output into the plugin binary. The NLDD
+// Design System is intentionally NOT bundled: the app uses <nldd-*> tags whose
+// registrations come from the shared /plugin-ui/nldd.js, loaded at runtime via
+// loadNldd(). See docs/funs/FUN-18.adoc.
 export default defineConfig({
   root,
   // Relative asset URLs so the built HTML resolves ./assets/* under /console/,
@@ -41,8 +41,8 @@ export default defineConfig({
     },
   },
   test: {
-    // The form logic in src/form.ts is pure DOM work (no NLDS runtime, no network),
-    // so a lightweight DOM is all it needs.
+    // The form logic in src/form.ts is pure DOM work (no NLDD Design System runtime,
+    // no network), so a lightweight DOM is all it needs.
     environment: 'happy-dom',
     include: ['src/**/*.test.ts'],
   },
